Moved noChests error to error folder.
