+++ src/errors/NoChestsError.ts
